+++ src/routes/events.js
@@ -11,7 +11,7 @@ router.use(bodyParser.json());
 router.use(bodyParser.urlencoded({ extended: false }));
 
 // authentication, authorization and preprocessing
-const { authenticateFromHeaderField } = require('../security/authentication');
+const { authenticateFromHeaderField, isMigration } = require('../security/authentication');
 const { authorizeAccessToScopeId, authorizeAccessToObjects, authorizeWithPotentialScopeIds } = require('../security/authorization');
 const jsonApiToJson = require('../parsers/event/jsonApiToJson');
 const icsToJson = require('../parsers/event/icsToJson');
@@ -26,7 +26,8 @@ const eventsToIcsInJsonApi = require('../parsers/event/eventsToIcsInJsonApi');
 const getEvents = require('../services/events/getEvents');
 const getOriginalEvent = require('../queries/original-events/getOriginalEvent');
 const insertEvents = require('../services/events/insertEvents');
-const deleteEventWithScope = require('../services/events/deleteEventWithScope');
+const { deleteEventWithScope } = require('../services/events/deleteEventWithScope');
+const { deleteDuplicatedEvents } = require('../services/events/deleteDuplicatedEvents');
 const updateEvents = require('../services/events/updateEvents');
 
 /* routes */
@@ -139,6 +140,26 @@ function sendUpdateNotification(updatedEvents) {
 	return updatedEvents;
 }
 
+router.delete('/events/duplicates', isMigration, (req, res, next) => {
+	deleteDuplicatedEvents()
+		.then((result) => ({data: result}))
+		.then((jsonApi) => { returnSuccess(res, 204, jsonApi); })
+		.catch(next);
+});
+
+const processDeletedEvents = (deletedEvents, res, next) => {
+	if (deletedEvents.length > 0) {
+		returnSuccess({data: deletedEvents}, 204);
+	} else {
+		const err = {
+			message: 'Given eventId or scopeIds not found for event deletion',
+			status: 404,
+			title: 'Query Error',
+		};
+		next(err);
+	}
+}
+
 router.delete('/events/:eventId', jsonApiToJson, authenticateFromHeaderField, function (req, res, next) {
 	const eventId = req.params.eventId;
 	const user = req.user;
@@ -147,24 +168,7 @@ router.delete('/events/:eventId', jsonApiToJson, authenticateFromHeaderField, fu
 	authorizeWithPotentialScopeIds(eventId, scopeIds, user, getEvents, getOriginalEvent, 'eventId')
 		.then(() => deleteEventWithScope(eventId, scopeIds))
 		.then((deletedEvents) => {
-			if (deletedEvents.length > 0) {
-				returnSuccess(res, 204);
-				deletedEvents.forEach((deletedEvent) => {
-					sendNotification.forDeletedEvent(
-						deletedEvent['scope_id'],
-						deletedEvent['summary'],
-						deletedEvent['dtstart'],
-						deletedEvent['dtend']
-					);
-				});
-			} else {
-				const err = {
-					message: 'Given eventId or scopeIds not found for event deletion',
-					status: 404,
-					title: 'Query Error',
-				};
-				next(err);
-			}
+			processDeletedEvents(deletedEvents, res, next);
 		})
 		.catch(next);
 });